Slimefun integration
This will solve some problems with Slimefun (also works with addons!). For example, if you start breaking a skull or block from Slimefun, stop and continue breaking, you will get this skull twice. I'm sure that players won't be angry that this system isn't applyed to skulls (skulls are easy to break ;) ) and blocks from Slimefun.